+++ webdev/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.1
+
+- Fix launching Chrome on Windows.
+
 ## 2.0.0
 
 ## Breaking Changes